pyproject.toml
[tool.poetry]
name = "actions-example-python"
version = "0.1.0"
description = "An example of GitHub Actions with Python"
authors = ["Brian Torres-Gil <brian@ixi.us>"]

[tool.poetry.dependencies]
python = "^3.6"
pan-os-python = "^1.0.1"
defusedxml = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"
flake8-bugbear = "^20.1.4"
mypy = "^0.782"
bandit = "^1.6.2"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
